Issue 299 not sending numpad commands correctly (#300)
* Fixes numpad numeric entry

* Updated SC screen templates

* Fixed shields for SC large tablet

* added null check to background path when changing background from color to image

* fixed display not updating after deleting a control when the background image is set

+++ gic_flutter/lib/src/views/screenEditor/backgroundDialog.dart
@@ -62,7 +62,7 @@ class _BackgroundDialogState extends State<BackgroundDialog> {
       Directory dest = await getApplicationDocumentsDirectory();
       String destPath = path.join(
           dest.path, "screens", screenViewModel.screenId.toString(), path.basename(file.path));
-      if (File(screenViewModel.backgroundPath).existsSync())
+      if (screenViewModel.backgroundPath != null && File(screenViewModel.backgroundPath).existsSync())
         File(screenViewModel.backgroundPath).deleteSync();
       File newFile = File(file.path).copySync(destPath);
       setState(() {
